dev-requirements.txt
sphinxcontrib.bibtex
pybtex
pytest
nbsphinx
Sphinx==5.3.0
sphinx_rtd_theme
numpydoc
docutils==0.19
sphinx-copybutton
IPython
nbsphinx-link
git+https://github.com/zincware/ZincHub.git
isort
coverage
coveralls
mdanalysis
black
znvis
ase
astroid
ruff>=0.0.231
